ENH: Fixes for return types and samples
- Fix for makefile that builds all libraries before testing
- Change AIORET_TYPE into long
- Fixed print statements that will print AIORET_TYPE statements
- Removed unnecessary continuous_setup call that is no longer needed
- Fixes for samples that print return codes

+++ AIOUSB/lib/AIOTypes.h
@@ -16,7 +16,7 @@
 #include <errno.h>
 
 
-typedef int AIORET_TYPE;        /* New return type is signed, negative indicates error */
+typedef long AIORET_TYPE;        /* New return type is signed, negative indicates error */
 typedef unsigned long AIORESULT;
 
 typedef unsigned short * COUNTS;
